requirements.txt
flask
python-telegram-bot
yt-dlp